backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
numpy
pymongo
PyJWT
bcrypt
email-validator
python-multipart
motor
certifi